Exp-1/output/f00069_AutoImageProcessor.from_pretrained_test.py
from f00069_AutoImageProcessor.from_pretrained import *
def test_from_pretrained():
    # Instantiate the image processor
    image_processor = AutoImageProcessor.from_pretrained("google/vit-base-patch16-224")

    # Test case 1
    assert image_processor is not None

    # Test case 2
    assert isinstance(image_processor, AutoImageProcessor)

    # Test case 3
    assert image_processor.config is not None

    # Test case 4
    assert image_processor.model is not None

    # Test case 5
    assert image_processor.tokenizer is not None


def test_auto_image_processor():
    test_from_pretrained()


test_auto_image_processor()
